requirements.txt
openai>=1.0.0
hypothesis>=6.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gTTS>=2.5.1
fastapi>=0.111.0
uvicorn>=0.30.0
gunicorn>=22.0.0
python-multipart>=0.0.9
pymongo>=4.0.0
